requirements.txt
pandas>=2.2.1
python-dotenv>=1.0.1
Requests>=2.31.0
tqdm>=4.66.4
